fix(chat): cast i64 error_msg_code to i32 for server_error_handler
The proto-generated ErrrMessage.error_msg_code is i64, but the
ServerErrorHandler callback expects i32. Add an explicit cast at the
call site to resolve the pre-existing type mismatch that blocked
building im-chat.

+++ im-chat/src/client.rs
@@ -560,7 +560,7 @@ impl ReadTask {
                                 );
                                 if let Some(ref handler) = self.server_error_handler {
                                     let handler = handler.clone();
-                                    let code = server_error.error_msg_code;
+                                    let code = server_error.error_msg_code as i32;
                                     let msg = server_error.error_msg.clone();
                                     tokio::spawn(async move {
                                         handler(code, msg).await;
